Module aws_sdk_networkfirewall::error

Modules

insufficient_capacity_exception

See InsufficientCapacityException

internal_server_error

See InternalServerError

invalid_operation_exception

See InvalidOperationException

invalid_request_exception

See InvalidRequestException

invalid_resource_policy_exception

See InvalidResourcePolicyException

invalid_token_exception

See InvalidTokenException

limit_exceeded_exception

See LimitExceededException

log_destination_permission_exception

See LogDestinationPermissionException

resource_not_found_exception

See ResourceNotFoundException

resource_owner_check_exception

See ResourceOwnerCheckException

throttling_exception

See ThrottlingException

unsupported_operation_exception

See UnsupportedOperationException

Structs

AssociateFirewallPolicyError

AssociateSubnetsError

CreateFirewallError

CreateFirewallPolicyError

CreateRuleGroupError

DeleteFirewallError

DeleteFirewallPolicyError

DeleteResourcePolicyError

DeleteRuleGroupError

DescribeFirewallError

DescribeFirewallPolicyError

DescribeLoggingConfigurationError

DescribeResourcePolicyError

DescribeRuleGroupError

DisassociateSubnetsError

InsufficientCapacityException

AWS doesn't currently have enough available capacity to fulfill your request. Try your

InternalServerError

Your request is valid, but Network Firewall couldn’t perform the operation because of a

InvalidOperationException

The operation failed because it's not valid. For example, you might have tried to delete

InvalidRequestException

The operation failed because of a problem with your request. Examples include:

InvalidResourcePolicyException

InvalidTokenException

The token you provided is stale or isn't valid for the operation.

LimitExceededException

Unable to perform the operation because doing so would violate a limit setting.

ListFirewallPoliciesError

ListFirewallsError

ListRuleGroupsError

ListTagsForResourceError

LogDestinationPermissionException

Unable to send logs to a configured logging destination.

PutResourcePolicyError

ResourceNotFoundException

Unable to locate a resource using the parameters that you provided.

ResourceOwnerCheckException

TagResourceError

ThrottlingException

Unable to process the request due to throttling limitations.

UnsupportedOperationException

The operation you requested isn't supported by Network Firewall.

UntagResourceError

UpdateFirewallDeleteProtectionError

UpdateFirewallDescriptionError

UpdateFirewallPolicyChangeProtectionError

UpdateFirewallPolicyError

UpdateLoggingConfigurationError

UpdateRuleGroupError

UpdateSubnetChangeProtectionError

Enums

AssociateFirewallPolicyErrorKind

AssociateSubnetsErrorKind

CreateFirewallErrorKind

CreateFirewallPolicyErrorKind

CreateRuleGroupErrorKind

DeleteFirewallErrorKind

DeleteFirewallPolicyErrorKind

DeleteResourcePolicyErrorKind

DeleteRuleGroupErrorKind

DescribeFirewallErrorKind

DescribeFirewallPolicyErrorKind

DescribeLoggingConfigurationErrorKind

DescribeResourcePolicyErrorKind

DescribeRuleGroupErrorKind

DisassociateSubnetsErrorKind

ListFirewallPoliciesErrorKind

ListFirewallsErrorKind

ListRuleGroupsErrorKind

ListTagsForResourceErrorKind

PutResourcePolicyErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateFirewallDeleteProtectionErrorKind

UpdateFirewallDescriptionErrorKind

UpdateFirewallPolicyChangeProtectionErrorKind

UpdateFirewallPolicyErrorKind

UpdateLoggingConfigurationErrorKind

UpdateRuleGroupErrorKind

UpdateSubnetChangeProtectionErrorKind