Module aws_sdk_networkfirewall::model

Modules

action_definition

See ActionDefinition

address

See Address

attachment

See Attachment

custom_action

See CustomAction

dimension

See Dimension

firewall

See Firewall

firewall_metadata

See FirewallMetadata

firewall_policy

See FirewallPolicy

firewall_policy_metadata

See FirewallPolicyMetadata

firewall_policy_response

See FirewallPolicyResponse

firewall_status

See FirewallStatus

header

See Header

ip_set

See IpSet

log_destination_config

See LogDestinationConfig

logging_configuration

See LoggingConfiguration

match_attributes

See MatchAttributes

per_object_status

See PerObjectStatus

port_range

See PortRange

port_set

See PortSet

publish_metric_action

See PublishMetricAction

rule_definition

See RuleDefinition

rule_group

See RuleGroup

rule_group_metadata

See RuleGroupMetadata

rule_group_response

See RuleGroupResponse

rule_option

See RuleOption

rule_variables

See RuleVariables

rules_source

See RulesSource

rules_source_list

See RulesSourceList

stateful_rule

See StatefulRule

stateful_rule_group_reference

See StatefulRuleGroupReference

stateless_rule

See StatelessRule

stateless_rule_group_reference

See StatelessRuleGroupReference

stateless_rules_and_custom_actions

See StatelessRulesAndCustomActions

subnet_mapping

See SubnetMapping

sync_state

See SyncState

tag

See Tag

tcp_flag_field

See TcpFlagField

Structs

ActionDefinition

A custom action to use in stateless rule actions settings. This is used in CustomAction.

Address

A single IP address specification. This is used in the MatchAttributes

Attachment

The configuration and status for a single subnet that you've specified for use by the

CustomAction

An optional, non-standard action to use for stateless packet handling. You can define

Dimension

The value to use in an Amazon CloudWatch custom metric dimension. This is used in the

Firewall

The firewall defines the configuration settings for an AWS Network Firewall firewall. These settings include the firewall policy, the subnets in your VPC to use for the firewall endpoints, and any tags that are attached to the firewall AWS resource.

FirewallMetadata

High-level information about a firewall, returned by operations like create and

FirewallPolicy

The firewall policy defines the behavior of a firewall using a collection of stateless

FirewallPolicyMetadata

High-level information about a firewall policy, returned by operations like create and

FirewallPolicyResponse

The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.

FirewallStatus

Detailed information about the current status of a Firewall. You can retrieve this for a firewall by calling DescribeFirewall and providing the firewall name and ARN.

Header

The 5-tuple criteria for AWS Network Firewall to use to inspect packet headers in stateful

IpSet

A list of IP addresses and address ranges, in CIDR notation. This is part of a RuleVariables.

LogDestinationConfig

Defines where AWS Network Firewall sends logs for the firewall for one log type. This is used

LoggingConfiguration

Defines how AWS Network Firewall performs logging for a Firewall.

MatchAttributes

Criteria for Network Firewall to use to inspect an individual packet in stateless rule inspection. Each match attributes set can include one or more items such as IP address, CIDR range, port number, protocol, and TCP flags.

PerObjectStatus

Provides configuration status for a single policy or rule group that is used for a firewall endpoint. Network Firewall

PortRange

A single port range specification. This is used for source and destination port ranges

PortSet

A set of port ranges for use in the rules in a rule group.

PublishMetricAction

Stateless inspection criteria that publishes the specified metrics to Amazon CloudWatch for the

RuleDefinition

The inspection criteria and action for a single stateless rule. AWS Network Firewall inspects each packet for the specified matching

RuleGroup

The object that defines the rules in a rule group. This, along with RuleGroupResponse, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

RuleGroupMetadata

High-level information about a rule group, returned by ListRuleGroups.

RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

RuleOption

Additional settings for a stateful rule. This is part of the StatefulRule configuration.

RuleVariables

Settings that are available for use in the rules in the RuleGroup

RulesSource

The stateless or stateful rules definitions for use in a single rule group. Each rule

RulesSourceList

Stateful inspection criteria for a domain list rule group.

StatefulRule

A single 5-tuple stateful rule, for use in a stateful rule group.

StatefulRuleGroupReference

Identifier for a single stateful rule group, used in a firewall policy to refer to a

StatelessRule

A single stateless rule. This is used in StatelessRulesAndCustomActions.

StatelessRuleGroupReference

Identifier for a single stateless rule group, used in a firewall policy to refer to the

StatelessRulesAndCustomActions

Stateless inspection criteria. Each stateless rule group uses exactly one of these data

SubnetMapping

The ID for a subnet that you want to associate with the firewall. This is used with

SyncState

The status of the firewall endpoint and firewall policy configuration for a single VPC

Tag

A key:value pair associated with an AWS resource. The key:value pair can be anything you

TcpFlagField

TCP flags and masks to inspect packets for, used in stateless rules MatchAttributes settings.

Enums

AttachmentStatus

ConfigurationSyncState

FirewallStatusValue

GeneratedRulesType

LogDestinationType

LogType

PerObjectSyncStatus

ResourceStatus

RuleGroupType

StatefulAction

StatefulRuleDirection

StatefulRuleProtocol

TargetType

TcpFlag