Module aws_sdk_networkfirewall::operation

Structs

AssociateFirewallPolicy

Associates a FirewallPolicy to a Firewall.

AssociateSubnets

Associates the specified subnets in the Amazon VPC to the firewall. You can specify one

CreateFirewall

Creates an AWS Network Firewall Firewall and accompanying FirewallStatus for a VPC.

CreateFirewallPolicy

Creates the firewall policy for the firewall according to the specifications.

CreateRuleGroup

Creates the specified stateless or stateful rule group, which includes the rules for

DeleteFirewall

Deletes the specified Firewall and its FirewallStatus. This operation requires the firewall's DeleteProtection flag to be

DeleteFirewallPolicy

Deletes the specified FirewallPolicy.

DeleteResourcePolicy

Deletes a resource policy that you created in a PutResourcePolicy request.

DeleteRuleGroup

Deletes the specified RuleGroup.

DescribeFirewall

Returns the data objects for the specified firewall.

DescribeFirewallPolicy

Returns the data objects for the specified firewall policy.

DescribeLoggingConfiguration

Returns the logging configuration for the specified firewall.

DescribeResourcePolicy

Retrieves a resource policy that you created in a PutResourcePolicy request.

DescribeRuleGroup

Returns the data objects for the specified rule group.

DisassociateSubnets

Removes the specified subnet associations from the firewall. This removes the

ListFirewallPolicies

Retrieves the metadata for the firewall policies that you have defined. Depending on

ListFirewalls

Retrieves the metadata for the firewalls that you have defined. If you provide VPC

ListRuleGroups

Retrieves the metadata for the rule groups that you have defined. Depending on your

ListTagsForResource

Retrieves the tags associated with the specified resource. Tags are key:value pairs that

PutResourcePolicy

Creates or updates an AWS Identity and Access Management policy for your rule group or firewall policy. Use this to share rule groups and firewall policies between accounts. This operation works in conjunction with the AWS Resource Access Manager (RAM) service

TagResource

Adds the specified tags to the specified resource. Tags are key:value pairs that you can

UntagResource

Removes the tags with the specified keys from the specified resource. Tags are key:value

UpdateFirewallDeleteProtection

Modifies the flag, DeleteProtection, which indicates whether it is possible

UpdateFirewallDescription

Modifies the description for the specified firewall. Use the description to help you

UpdateFirewallPolicy

Updates the properties of the specified firewall policy.

UpdateFirewallPolicyChangeProtection

UpdateLoggingConfiguration

Sets the logging configuration for the specified firewall.

UpdateRuleGroup

Updates the rule settings for the specified rule group. You use a rule group by

UpdateSubnetChangeProtection